moderation-service/pyproject.toml
[project]
name = "moderation-service"
version = "0.1.0"
description = "Moderation Service for Panchayat"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "aiokafka>=0.10.0",
    "alt-profanity-check>=1.4.1.post1",
    "httpx>=0.28.1",
    "redis>=7.4.0",
    "protobuf>=5.0.0",
    "opentelemetry-exporter-otlp>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
]
